backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
minio==7.2.0
cryptography==41.0.8
httpx==0.25.2
docker==6.1.3
aiofiles==23.2.1
python-dotenv==1.0.0